pet_store/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
python-dotenv==0.19.1
Werkzeug==2.0.1
Flask-Session==0.3.2
